Extract the common logic for upsert metadata manager
